refactor(format): move Style, PaperSize and the remaining formatters
